Implement DAWG.__sizeof__
This gives a better approximation of the memory used by a DAWG object
instance by adding the memory size of the dictionary.

It could be further improved by including the size of the DAWG itself,
but I wasn't able to find a nice way to do that with the current APIs.